+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- improved `sourceFileMap` resolution
+
 ## [4.0.17] - 2021-12-27
 
 ### Changed